Fix the way headings (angles between points) are stored and processed
A heading is a property shared by two neighboring points, so it doesn't
make sense to store this on a point object itself.  Store data as a
TurtlePolygon instead, that knows its own Points and Headings.

The Headings array is redundant with the point coordinates but this is
stored to avoid a lot of recalculation of the angles between points.